requirements.txt
Django
djangorestframework
django-filter
djangorestframework-simplejwt
drf-spectacular
psycopg2-binary
django-cors-headers
python-decouple
